# client/openapi/lumen.openapi.json
{
  "components": {
    "schemas": {
      "CycleDayLogResponse": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "day": {
            "format": "date",
            "type": "string"
          },
          "mood": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "notes": {
            "nullable": true,
            "type": "string"
          },
          "pain": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "CycleDayResponse": {
        "additionalProperties": false,
        "properties": {
          "date": {
            "format": "date",
            "type": "string"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/CycleEventResponse"
            },
            "nullable": true,
            "type": "array"
          },
          "log": {
            "$ref": "#/components/schemas/CycleDayLogResponse"
          },
          "phaseOverrides": {
            "items": {
              "$ref": "#/components/schemas/PhaseOverrideBoundary"
            },
            "nullable": true,
            "type": "array"
          }
        },
        "type": "object"
      },
      "CycleEventResponse": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "flowIntensity": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "kind": {
            "nullable": true,
            "type": "string"
          },
          "notes": {
            "nullable": true,
            "type": "string"
          },
          "occurredOn": {
            "format": "date",
            "type": "string"
          },
          "source": {
            "nullable": true,
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "HttpValidationProblemDetails": {
        "additionalProperties": {},
        "properties": {
          "detail": {
            "nullable": true,
            "type": "string"
          },
          "errors": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "nullable": true,
            "type": "object"
          },
          "instance": {
            "nullable": true,
            "type": "string"
          },
          "status": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "title": {
            "nullable": true,
            "type": "string"
          },
          "type": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
      "LogCycleDayRequest": {
        "additionalProperties": false,
        "properties": {
          "mood": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "notes": {
            "nullable": true,
            "type": "string"
          },
          "pain": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          }
        },
        "type": "object"
      },
      "LogCycleEventRequest": {
        "additionalProperties": false,
        "properties": {
          "flowIntensity": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "kind": {
            "nullable": true,
            "type": "string"
          },
          "notes": {
            "nullable": true,
            "type": "string"
          },
          "occurredOn": {
            "format": "date",
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
      "MeResponse": {
        "additionalProperties": false,
        "properties": {
          "displayName": {
            "nullable": true,
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "locale": {
            "nullable": true,
            "type": "string"
          },
          "onboardingCompleted": {
            "type": "boolean"
          },
          "timezone": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
      "OnboardingStartRequest": {
        "additionalProperties": false,
        "properties": {
          "displayName": {
            "nullable": true,
            "type": "string"
          },
          "email": {
            "nullable": true,
            "type": "string"
          },
          "locale": {
            "nullable": true,
            "type": "string"
          },
          "password": {
            "nullable": true,
            "type": "string"
          },
          "policyVersion": {
            "nullable": true,
            "type": "string"
          },
          "timezone": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
      "OnboardingStartResponse": {
        "additionalProperties": false,
        "properties": {
          "userId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "type": "object"
      },
      "PhaseOverrideBoundary": {
        "additionalProperties": false,
        "properties": {
          "boundary": {
            "nullable": true,
            "type": "string"
          },
          "occurredOn": {
            "format": "date",
            "type": "string"
          },
          "phase": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
      "PhaseOverrideInput": {
        "additionalProperties": false,
        "properties": {
          "boundary": {
            "nullable": true,
            "type": "string"
          },
          "occurredOn": {
            "format": "date",
            "nullable": true,
            "type": "string"
          },
          "phase": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
      "PhaseOverridesResponse": {
        "additionalProperties": false,
        "properties": {
          "boundaries": {
            "items": {
              "$ref": "#/components/schemas/PhaseOverrideBoundary"
            },
            "nullable": true,
            "type": "array"
          },
          "cycleStartOn": {
            "format": "date",
            "type": "string"
          }
        },
        "type": "object"
      },
      "ProblemDetails": {
        "additionalProperties": {},
        "properties": {
          "detail": {
            "nullable": true,
            "type": "string"
          },
          "instance": {
            "nullable": true,
            "type": "string"
          },
          "status": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "title": {
            "nullable": true,
            "type": "string"
          },
          "type": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
      "QuickCheckinRequest": {
        "additionalProperties": false,
        "properties": {
          "mood": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "pain": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          }
        },
        "type": "object"
      },
      "QuickCheckinResponse": {
        "additionalProperties": false,
        "properties": {
          "day": {
            "format": "date",
            "type": "string"
          },
          "mood": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "pain": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "SavePhaseOverridesRequest": {
        "additionalProperties": false,
        "properties": {
          "boundaries": {
            "items": {
              "$ref": "#/components/schemas/PhaseOverrideInput"
            },
            "nullable": true,
            "type": "array"
          },
          "cycleStartOn": {
            "format": "date",
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
      "UpdateMeRequest": {
        "additionalProperties": false,
        "properties": {
          "displayName": {
            "nullable": true,
            "type": "string"
          },
          "locale": {
            "nullable": true,
            "type": "string"
          },
          "timezone": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Lumen.Api",
    "version": "1.0"
  },
  "openapi": "3.0.4",
  "paths": {
    "/checkin/quick": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickCheckinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuickCheckinResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "tags": [
          "Lumen.Api"
        ]
      }
    },
    "/cycle/day/{date}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "date",
            "required": true,
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CycleDayResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "tags": [
          "Lumen.Api"
        ]
      },
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "date",
            "required": true,
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogCycleDayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CycleDayLogResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "tags": [
          "Lumen.Api"
        ]
      }
    },
    "/cycle/events": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogCycleEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CycleEventResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "tags": [
          "Lumen.Api"
        ]
      }
    },
    "/cycle/events/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "tags": [
          "Lumen.Api"
        ]
      }
    },
    "/cycle/phase-override": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavePhaseOverridesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhaseOverridesResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "tags": [
          "Lumen.Api"
        ]
      }
    },
    "/health": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Lumen.Api"
        ]
      }
    },
    "/health/ready": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Lumen.Api"
        ]
      }
    },
    "/me": {
      "delete": {
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Lumen.Api"
        ]
      },
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "tags": [
          "Lumen.Api"
        ]
      },
      "patch": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "tags": [
          "Lumen.Api"
        ]
      }
    },
    "/onboarding/start": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingStartResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          }
        },
        "tags": [
          "Lumen.Api"
        ]
      }
    }
  },
  "tags": [
    {
      "name": "Lumen.Api"
    }
  ]
}